// packages/plugin/src/components/resizer/types.ts
export type Positioning = {
  currentHeight: "inherit" | number;
  currentWidth: "inherit" | number;
  direction: number;
  isResizing: boolean;
  ratio: number;
  startHeight: number;
  startWidth: number;
  startX: number;
  startY: number;
};

export type UserSelect = {
  priority: string;
  value: string;
};